moved activities to correct section
