+++ app/src/main/java/Activities/ViewUserActivity.java
@@ -1,6 +1,9 @@
-package com.cmput301f18t20.medicalphotorecord;
+package Activities;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.cmput301f18t20.medicalphotorecord.R;
+
 public class ViewUserActivity extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
