+++ app/src/main/res/layout/activity_browse_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BrowseUserActivity">
+    tools:context="Activities.BrowseUserActivity">
      <ListView
         android:id="@+id/browse_user_id"
         android:layout_width="match_parent"
