+++ app/src/main/res/layout/activity_view_user.xml
@@ -4,5 +4,5 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewUserActivity">
+    tools:context="Activities.ViewUserActivity">
  </android.support.constraint.ConstraintLayout>
